% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{changeBase}
\alias{changeBase}
\title{Translate base slot based on complementary bases dictionary.
Base sequence in transformed using compl_target column.}
\usage{
changeBase(compl_dict, bases)
}
\arguments{
\item{compl_dict}{complementary bases dictionary}

\item{bases}{string, one or two-elements vector}
}
\value{
string
}
\description{
Translate base slot based on complementary bases dictionary.
Base sequence in transformed using compl_target column.
}
