% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{complementary_bases}
\alias{complementary_bases}
\title{Default XNAString complementarity dictionary}
\format{
A data.table with 6 rows and 3 variables:
\describe{
\item{base}{base symbol}
\item{target}{complementary base}
\item{compl_target}{complementary target}
}
}
\source{
RMR internal bioinformatics database (Mimir)
}
\usage{
data(complementary_bases)
}
\description{
A dataset containing default internal XNAString dictionary
with base complemetary.
}
\keyword{datasets}
