% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setterGetter.R
\name{default_sugar}
\alias{default_sugar}
\alias{default_sugar,XNAString-method}
\alias{default_sugar,XNAStringSet-method}
\alias{default_sugar<-}
\alias{default_sugar<-,XNAString-method}
\title{Default_sugar setter/getter method}
\usage{
default_sugar(x, ...)

\S4method{default_sugar}{XNAString}(x)

\S4method{default_sugar}{XNAStringSet}(x)

default_sugar(x, ...) <- value

\S4method{default_sugar}{XNAString}(x) <- value
}
\arguments{
\item{x}{XNAString/XNAStringSet object}

\item{...}{optional arguments to generic function to support
additional methods}

\item{value}{character vector applied only for setter method}
}
\value{
vector in getter method, XNAStringSet object
(with replaced name slot) in setter method
}
\description{
Getter methods enable extraction of single slots from XNAString
and XNAStringSet objects.
E.g. name method extracts name slot from XNAString/XNAStringSet
object.
}
\details{
Setter methods enable overwriting single slots from XNAString
and XNAStringSet objects.
E.g. name<- method overwrites existing name slot
}
\examples{

my_dic <- data.table::data.table(
  type = c(
    rep("base", 3),
    rep("sugar", 2),
    rep("backbone", 3)
  ),
  symbol = c("G", "E", "A", "F", "O", "S", "B", "X")
)
obj <- XNAString(
  name = "b",
  base = "GGE",
  default_sugar = 'F',
  default_backbone = 'X',
  dictionary = my_dic
)
default_sugar(obj)
}
