% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setterGetter.R
\name{dictionary}
\alias{dictionary}
\alias{dictionary,XNAString-method}
\alias{dictionary<-}
\alias{dictionary<-,XNAString-method}
\title{Dictionary setter/getter method}
\usage{
dictionary(x, ...)

\S4method{dictionary}{XNAString}(x)

dictionary(x, ...) <- value

\S4method{dictionary}{XNAString}(x) <- value
}
\arguments{
\item{x}{XNAString/XNAStringSet object}

\item{...}{optional arguments to generic function to support additional methods}

\item{value}{character vector applied only for setter method}
}
\value{
vector in getter method, XNAStringSet object (with replaced name slot) in setter method
}
\description{
Getter methods enable extraction of single slots from XNAString
and XNAStringSet objects.
E.g. name method extracts name slot from XNAString/XNAStringSet
object.
}
\details{
Setter methods enable overwriting single slots from XNAString
and XNAStringSet objects.
E.g. name<- method overwrites existing name slot
}
