% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDuplexStructure.R
\name{predictDuplexStructure}
\alias{predictDuplexStructure}
\alias{predictDuplexStructureFun}
\alias{predictDuplexStructure,XNAString-method}
\title{Compute Minimum Free Energy (MFE), and a corresponding secondary structure
for two dimerized RNA sequences.}
\usage{
predictDuplexStructureFun(obj)

predictDuplexStructure(obj, ...)

\S4method{predictDuplexStructure}{XNAString}(obj)
}
\arguments{
\item{obj}{XNAString object}

\item{...}{optional arguments to generic function to support
additional methods}
}
\value{
list (structure and mfe)
}
\description{
This function is a wrapper for RNAcofold from ViennaRNA package.
}
\examples{
obj1 <- XNAString(
  base = "ATCG",
  sugar = "FODD",
  conjugate3 = "TAG"
)
predictDuplexStructure(obj1)
}
