% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseComplement.R
\name{generateAllCombinations}
\alias{generateAllCombinations}
\title{Generate all string combinations based on IUPAC dictionary}
\usage{
generateAllCombinations(str, iupac_dict)
}
\arguments{
\item{str}{character string of bases with IUPAC symbols}

\item{iupac_dict}{data.table mapping IUPAC symbols to bases}
}
\description{
Generate all string combinations based on IUPAC dictionary
}
