% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setterGetter.R
\docType{methods}
\name{name}
\alias{name}
\alias{getterMethods}
\alias{name,XNAString-method}
\alias{XNAString}
\alias{name,XNAStringSet-method}
\alias{XNAStringSet}
\alias{base}
\alias{base,XNAString-method}
\alias{base,XNAStringSet-method}
\alias{sugar}
\alias{sugar,XNAString-method}
\alias{sugar,XNAStringSet-method}
\alias{backbone}
\alias{backbone,XNAString-method}
\alias{backbone,XNAStringSet-method}
\alias{target}
\alias{target,XNAString-method}
\alias{target,XNAStringSet-method}
\alias{conjugate5}
\alias{conjugate5,XNAString-method}
\alias{conjugate5,XNAStringSet-method}
\alias{conjugate3}
\alias{conjugate3,XNAString-method}
\alias{conjugate3,XNAStringSet-method}
\alias{dictionary}
\alias{dictionary,XNAString-method}
\alias{secondary_structure}
\alias{secondary_structure,XNAString-method}
\alias{secondary_structure,XNAStringSet-method}
\alias{compl_dictionary}
\alias{compl_dictionary,XNAString-method}
\alias{duplex_structure}
\alias{duplex_structure,XNAString-method}
\alias{duplex_structure,XNAStringSet-method}
\title{Set of getter methods}
\usage{
name(x, ...)

\S4method{name}{XNAString}(x)

\S4method{name}{XNAStringSet}(x, i = 1)

base(x, ...)

\S4method{base}{XNAString}(x)

\S4method{base}{XNAStringSet}(x, i = 1)

sugar(x, ...)

\S4method{sugar}{XNAString}(x)

\S4method{sugar}{XNAStringSet}(x, i = 1)

backbone(x, ...)

\S4method{backbone}{XNAString}(x)

\S4method{backbone}{XNAStringSet}(x, i = 1)

target(x, ...)

\S4method{target}{XNAString}(x)

\S4method{target}{XNAStringSet}(x, i = 1)

conjugate5(x, ...)

\S4method{conjugate5}{XNAString}(x)

\S4method{conjugate5}{XNAStringSet}(x, i = 1)

conjugate3(x, ...)

\S4method{conjugate3}{XNAString}(x)

\S4method{conjugate3}{XNAStringSet}(x, i = 1)

dictionary(x, ...)

\S4method{dictionary}{XNAString}(x)

secondary_structure(x, ...)

\S4method{secondary_structure}{XNAString}(x)

\S4method{secondary_structure}{XNAStringSet}(x)

compl_dictionary(x, ...)

\S4method{compl_dictionary}{XNAString}(x)

duplex_structure(x, ...)

\S4method{duplex_structure}{XNAString}(x)

\S4method{duplex_structure}{XNAStringSet}(x)
}
\arguments{
\item{x}{XNAString/XNAStringSet object}

\item{...}{optional arguments to generic function to support additional methods}

\item{i}{numeric - possibilities: 1 or 2. If 1 - 1st slots elements printed
out, 2nd otherwise. In case the second element is not in the object, empty
char created. This parameter is only available for XNAStringSet objects.}
}
\value{
vector
}
\description{
Getter methods enable extraction of single slots from XNAString
and XNAStringSet objects.
E.g. name method extracts name slot from XNAString/XNAStringSet
object.
}
\examples{

my_dic <- data.table::data.table(type = c(rep('base',3),
                                          rep('sugar',2),
                                          rep('backbone',3)),
                               symbol = c('G', 'E', 'A', 'F', 'O', 'S', 'B', 'X'))
obj <- XNAString(name = 'b',
                  base = 'GGE',
                  sugar = 'FFO',
                  dictionary = my_dic)
name(obj)
my_dic <- data.table::data.table(type = c(rep('base',3),
                                          rep('sugar',2),
                                          rep('backbone',3)),
                               symbol = c('G', 'E', 'A', 'F', 'O', 'S', 'B', 'X'))
obj1 <- XNAString(name = 'b',
                  base = 'GGE',
                  sugar = 'FFO',
                  dictionary = my_dic)
obj2 <- XNAString(name = 'b',
                  base = c('GGE','EEE'),
                  sugar = c('FFO', 'OOO'),
                  dictionary = my_dic)
XNAStringSetObj <- XNAStringSet(objects=list(obj1, obj2))
name(XNAStringSetObj)
}
