% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{instanceOf}
\alias{instanceOf}
\title{Check on an object type}
\usage{
instanceOf(object, type)
}
\arguments{
\item{object}{an object of any class}

\item{type}{class of an object}
}
\value{
logical information. TRUE if \code{object} class equals \code{type}
}
\description{
Check on an object type
}
\examples{
instanceOf(1, 'numeric')

}
