% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listOflists2Dt}
\alias{listOflists2Dt}
\title{Save list of lists as data.table}
\usage{
listOflists2Dt(list_of_lists)
}
\arguments{
\item{list_of_lists}{list of lists that will be saved as data.table.}
}
\value{
data.table
}
\description{
Save list of lists as data.table
}
\examples{
nested_list <- list(
  list(base = c("T"), sugar = c("G")),
  list(base = c("U"), sugar = c("G"))
)
listOflists2Dt(nested_list)
}
