% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictMfeStructure.R
\name{predictMfeStructure}
\alias{predictMfeStructure}
\alias{predictMfeStructureFun}
\alias{predictMfeStructure,XNAString-method}
\title{Prediction of MFE structure with ViennaRNA package}
\usage{
predictMfeStructureFun(obj)

predictMfeStructure(obj, ...)

\S4method{predictMfeStructure}{XNAString}(obj)
}
\arguments{
\item{obj}{XNAString object}

\item{...}{optional arguments to generic function to support additional methods}
}
\description{
This function is a wrapper for RNAfold from ViennaRNA package.
}
\examples{
 obj1 <- XNAString(base = 'ATCG',
                   sugar = 'FODD',
                   conjugate3 = 'TAG')
 predictMfeStructure(obj1) 
}
