% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helm2String.R
\name{parseRnaHelmComponent}
\alias{parseRnaHelmComponent}
\title{Parse monomers from HELM to multi-string notation}
\usage{
parseRnaHelmComponent(rna_component, dictionary = xna_dictionary)
}
\arguments{
\item{rna_component}{list of monomers building RNA}

\item{dictionary}{data.table with following columns:
"HELM", "type", "symbol".
By default internal XNAString dictionary is used.}
}
\value{
list of three strings: base, sugar, backbone
}
\description{
Parse monomers from HELM to multi-string notation
}
\examples{
parseRnaHelmComponent(c("[dR](A)P","[dR](A)P","[dR](A)"))
}
\author{
Marianna Plucinska
}
