% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceBase.R
\name{replaceBase}
\alias{replaceBase}
\title{Fucntion for replacing bases in XNAString object}
\usage{
replaceBase(xnastring_obj, base_to_replace, base_replacement, sugar)
}
\arguments{
\item{xnastring_obj}{XNAString object}

\item{base_to_replace}{string of bases that should be replaced}

\item{base_replacement}{string of bases that will be introduced}

\item{sugar}{string of sugars for which base will be replaced}
}
\value{
XNADtring object
}
\description{
replaceBase function replace one base for another
in positions with specified sugar
}
