% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseComplement.R
\name{reverseComplementFun}
\alias{reverseComplementFun}
\title{Reverse complement sequence based on dictionary}
\usage{
reverseComplementFun(obj)
}
\arguments{
\item{obj}{XNAString object}
}
\value{
string with reverse complement sequence
}
\description{
Reverse complement sequence based on dictionary
}
