% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabetFrequency.R
\name{seqAlphabetFrequency}
\alias{seqAlphabetFrequency}
\title{Create set of functions and methods to calculate alphabet frequency in
base, sugar and backbone slots}
\usage{
seqAlphabetFrequency(unique_letters, seq, as.prob)
}
\arguments{
\item{unique_letters}{string (or character) - these letters pose column names}

\item{seq}{string (or character) - frequency is calculated for
this string}

\item{as.prob}{logical - if TRUE frequency returned as probability of
occurrence}
}
\value{
numeric - named numeric vector
}
\description{
Create set of functions and methods to calculate alphabet frequency in
base, sugar and backbone slots
}
\examples{
seqAlphabetFrequency(c("A", "B", "C"), c("AABA"), as.prob = FALSE)
}
