% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphabetFrequency.R
\name{seqVectorAlphabetFrequency}
\alias{seqVectorAlphabetFrequency}
\title{seqVectorAlphabetFrequency function calculates frequency for strings vector}
\usage{
seqVectorAlphabetFrequency(unique_letters, seq_vec, as.prob)
}
\arguments{
\item{unique_letters}{string (or character) - these letters pose column names}

\item{seq_vec}{vector of strings (or characters) - frequency will be
calculated for this vector}

\item{as.prob}{logical - if TRUE frequency returned as probability of
occurence}
}
\value{
matrix - each row denotes frequency for a specific string of vector
}
\description{
seqVectorAlphabetFrequency function calculates frequency for strings vector
}
\examples{
seqVectorAlphabetFrequency(c('A', 'B', 'C'),
                                     c('AABA', 'BBBCCC'),
                                     as.prob=FALSE)

}
