% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinucleotideFrequency.R
\name{seqVectorDinucleotideFrequency}
\alias{seqVectorDinucleotideFrequency}
\title{seqVectorDinucleotideFrequency function calculates frequency for strings vector}
\usage{
seqVectorDinucleotideFrequency(unique_sets, seq_vec, as.prob)
}
\arguments{
\item{unique_sets}{string vector of double letters -these letters pose column names}

\item{seq_vec}{vector of strings (or characters) - frequency will be
calculated for this vector}

\item{as.prob}{logical - if TRUE frequency returned as probability of
occurence}
}
\value{
matrix - each row denotes frequency for a specific string of vector
}
\description{
seqVectorDinucleotideFrequency function calculates frequency for strings vector
}
\examples{
seqVectorDinucleotideFrequency(c("AB", "BA", "CD"), c("ABABAB", "ABABCD"), as.prob = FALSE)
}
