% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setterGetter.R
\docType{methods}
\name{name<-}
\alias{name<-}
\alias{setterMethods}
\alias{name}
\alias{name<-,XNAString-method}
\alias{XNAString}
\alias{name<-,XNAStringSet-method}
\alias{XNAStringSet}
\alias{base<-}
\alias{base}
\alias{base<-,XNAString-method}
\alias{base<-,XNAStringSet-method}
\alias{sugar<-}
\alias{sugar}
\alias{sugar<-,XNAString-method}
\alias{sugar<-,XNAStringSet-method}
\alias{backbone<-}
\alias{backbone}
\alias{backbone<-,XNAString-method}
\alias{backbone<-,XNAStringSet-method}
\alias{target<-}
\alias{target}
\alias{target<-,XNAString-method}
\alias{target<-,XNAStringSet-method}
\alias{conjugate5<-}
\alias{conjugate5}
\alias{conjugate5<-,XNAString-method}
\alias{conjugate5<-,XNAStringSet-method}
\alias{conjugate3<-}
\alias{conjugate3}
\alias{conjugate3<-,XNAString-method}
\alias{conjugate3<-,XNAStringSet-method}
\alias{dictionary<-}
\alias{dictionary}
\alias{dictionary<-,XNAString-method}
\alias{compl_dictionary<-}
\alias{compl_dictionary}
\alias{compl_dictionary<-,XNAString-method}
\alias{secondary_structure<-}
\alias{secondary_structure}
\alias{secondary_structure<-,XNAString-method}
\alias{duplex_structure<-}
\alias{duplex_structure}
\alias{duplex_structure<-,XNAString-method}
\title{Set of setter methods}
\usage{
name(x, ...) <- value

\S4method{name}{XNAString}(x) <- value

\S4method{name}{XNAStringSet}(x, i = 1) <- value

base(x, ...) <- value

\S4method{base}{XNAString}(x) <- value

\S4method{base}{XNAStringSet}(x, i = 1) <- value

sugar(x, ...) <- value

\S4method{sugar}{XNAString}(x) <- value

\S4method{sugar}{XNAStringSet}(x, i = 1) <- value

backbone(x, ...) <- value

\S4method{backbone}{XNAString}(x) <- value

\S4method{backbone}{XNAStringSet}(x, i = 1) <- value

target(x, ...) <- value

\S4method{target}{XNAString}(x) <- value

\S4method{target}{XNAStringSet}(x, i = 1) <- value

conjugate5(x, ...) <- value

\S4method{conjugate5}{XNAString}(x) <- value

\S4method{conjugate5}{XNAStringSet}(x, i = 1) <- value

conjugate3(x, ...) <- value

\S4method{conjugate3}{XNAString}(x) <- value

\S4method{conjugate3}{XNAStringSet}(x, i = 1) <- value

dictionary(x, ...) <- value

\S4method{dictionary}{XNAString}(x) <- value

compl_dictionary(x, ...) <- value

\S4method{compl_dictionary}{XNAString}(x) <- value

secondary_structure(x, ...) <- value

\S4method{secondary_structure}{XNAString}(x) <- value

duplex_structure(x, ...) <- value

\S4method{duplex_structure}{XNAString}(x) <- value
}
\arguments{
\item{x}{XNAString/XNAStringSet object}

\item{value}{string vector}

\item{...}{optional arguments to generic function to support additional methods}

\item{i}{numeric - possibilities: 1 or 2. If 1 - 1st slots elements printed
out, 2nd otherwise. In case the second element is not in the object, empty
char created. This parameter is only available for XNAStringSet objects.}
}
\value{
XNAStringSet object (with replaced name slot)
}
\description{
Setter methods enable overwriting single slots from XNAString
and XNAStringSet objects.
E.g. name<- method overwrites existing name slot
}
\examples{
my_dic <- data.table::data.table(type = c(rep('base',3),
                                          rep('sugar',2),
                                          rep('backbone',3)),
                               symbol = c('G', 'E', 'A', 'F', 'O', 'S', 'B', 'X'))
obj <- XNAString(name = 'b',
                  base = 'GGE',
                  sugar = 'FFO',
                  dictionary = my_dic)
name(obj) <- 'new_name'
my_dic <- data.table::data.table(type = c(rep('base',3),
                                          rep('sugar',2),
                                          rep('backbone',3)),
                               symbol = c('G', 'E', 'A', 'F', 'O', 'S', 'B', 'X'))
obj1 <- XNAString(name = 'b',
                  base = 'GGE',
                  sugar = 'FFO',
                  dictionary = my_dic)
obj2 <- XNAString(name = 'b',
                  base = c('GGE','EEE'),
                  sugar = c('FFO', 'OOO'),
                  dictionary = my_dic)
XNAStringSetObj <- XNAStringSet(objects=list(obj1, obj2))
name(XNAStringSetObj) <- c('new1', 'new2')
}
