% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XNAString2Helm.R
\name{siRNA_HELM}
\alias{siRNA_HELM}
\title{siRNA_HELM function takes XNAString object and returns pairing information
for base slot. Works only for double stranded molecules.}
\usage{
siRNA_HELM(xnastring_obj)
}
\arguments{
\item{xnastring_obj}{XNAString object}
}
\value{
string
}
\description{
siRNA_HELM function takes XNAString object and returns pairing information
for base slot. Works only for double stranded molecules.
}
\examples{
 obj1 <- XNAString(base =  c("CCCCUGCCGUGGUUCAUAA", "UUAUGAACCACGGCAGGGGCG"), 
                   sugar = c("OOFOFOFOFOFOFOFOFOF", "FFOFOFOFOFOFOFOFOFOFO"),
                   backbone = c("OOOOOOOOOOOOOOOOOO", "OOOOOOOOOOOOOOOOOOOO"),
                   conjugate3 = c(''))
                   
 siRNA_HELM(obj1)
 
}
