% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setterGetter.R
\name{sugar}
\alias{sugar}
\alias{sugar,XNAString-method}
\alias{sugar,XNAStringSet-method}
\alias{sugar<-}
\alias{sugar<-,XNAString-method}
\alias{sugar<-,XNAStringSet-method}
\title{Sugar setter/getter method}
\usage{
sugar(x, ...)

\S4method{sugar}{XNAString}(x)

\S4method{sugar}{XNAStringSet}(x, i = 1)

sugar(x, ...) <- value

\S4method{sugar}{XNAString}(x) <- value

\S4method{sugar}{XNAStringSet}(x, i = 1) <- value
}
\arguments{
\item{x}{XNAString/XNAStringSet object}

\item{...}{optional arguments to generic function to support additional methods}

\item{i}{numeric - possibilities: 1 or 2. If 1 - 1st slots elements printed
out, 2nd otherwise. In case the second element is not in the object, empty
char created. This parameter is only available for XNAStringSet objects.}

\item{value}{character vector applied only for setter method}
}
\value{
vector in getter method, XNAStringSet object (with replaced name slot) in setter method
}
\description{
Getter methods enable extraction of single slots from XNAString
and XNAStringSet objects.
E.g. name method extracts name slot from XNAString/XNAStringSet
object.
}
\details{
Setter methods enable overwriting single slots from XNAString
and XNAStringSet objects.
E.g. name<- method overwrites existing name slot
}
