% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uniqueChars}
\alias{uniqueChars}
\title{Utility functions useful when programming and developing XNAString class}
\usage{
uniqueChars(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
A list of vectors with unique characters found in \code{x} string
}
\description{
Utility functions useful when programming and developing XNAString class
}
\examples{
uniqueChars("TRGFFTR")
uniqueChars(c("TRGFFTR", "AATGRC"))
}
