% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xna_dictionary}
\alias{xna_dictionary}
\title{Default XNAString dictionary}
\format{
A data.table with 20 rows and 3 variables:
\describe{
\item{HELM}{HELM sequence coding monomer}
\item{type}{if element is coding base, sugar, backbone}
\item{symbol}{single string translation of HELM}
}
}
\source{
RMR internal bioinformatics database (Mimir)
}
\usage{
data(xna_dictionary)
}
\description{
A dataset containing default internal XNAString dictionary
with HELM to string translation.
}
\keyword{datasets}
