% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnaStringSetClass.R
\name{xnaObj2Dt}
\alias{xnaObj2Dt}
\title{xnaObj2Dt function - changes XNAString object to data.table}
\usage{
xnaObj2Dt(obj, slots)
}
\arguments{
\item{obj}{XNAString object}

\item{slots}{slots that are saved as column names (possibilities: "name",
"base", "sugar", "backbone", "target", "conjugate5",
"conjugate3" and  "dictionary" )}
}
\value{
data.table
}
\description{
xnaObj2Dt function - changes XNAString object to data.table
}
