% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnaStringSetClass.R
\name{XNAString2XNAStringSet}
\alias{XNAString2XNAStringSet}
\title{XNAString2XNAStringSet function - changes XNAString object to XNAStringSet}
\usage{
XNAString2XNAStringSet(XNAString_obj)
}
\arguments{
\item{XNAString_obj}{XNAString object}
}
\value{
XNAStringSet object
}
\description{
XNAString2XNAStringSet function - changes XNAString object to XNAStringSet
}
