% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classUnion.R
\docType{class}
\name{xnastringClassUnions}
\alias{xnastringClassUnions}
\alias{charOrDNA-class}
\title{setClassUnion definitions}
\description{
setClassUnion definitions used in XNAString class.
charOrDNAOrRNA consists of character, DNAString, RNAString, DNAStringSet,
RNAStringSet.
charOrDNA consists of character, DNAString, DNAStringSet
}
