% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnaStringSetClass.R
\docType{class}
\name{XNAStringSet-class}
\alias{XNAStringSet-class}
\alias{XNAStringSetMethod}
\alias{XNAStringSet}
\alias{show,XNAStringSet-method}
\alias{[,XNAStringSet,ANY,ANY,ANY-method}
\alias{extractionMethods}
\alias{[[,XNAStringSet,ANY,ANY-method}
\title{Create class which consists of XNAString objects given as a list}
\usage{
XNAStringSet(objects)

\S4method{show}{XNAStringSet}(object)

\S4method{[}{XNAStringSet,ANY,ANY,ANY}(x, i)

\S4method{[[}{XNAStringSet,ANY,ANY}(x, i)
}
\arguments{
\item{objects}{list of XNAString objects}

\item{object}{XNAStringSet object}

\item{x}{XNAStringSet object}

\item{i}{numeric, integer, character, logical - filter needed for extraction
method}
}
\value{
XNASTringSet object
}
\description{
Create class which consists of XNAString objects given as a list

Define show method

Method to extract a row/rows (either by row index or by 'name' slot)
XNAStringSet object is returned.

Method to extract a single row (either by row index or by 'name' slot)
XNAString object is returned.
}
\examples{
my_dic <- data.table::data.table(
  type = c(
    rep("base", 3),
    rep("sugar", 2),
    rep("backbone", 3)
  ),
  symbol = c("G", "E", "A", "F", "O", "S", "B", "X")
)
obj1 <- XNAString(
  name = "a",
  base = "GGE",
  sugar = "FFO",
  backbone = "SB",
  dictionary = my_dic
)
obj2 <- XNAString(
  name = "b",
  base = "GGE",
  sugar = "FFO",
  dictionary = my_dic
)
obj3 <- XNAString(
  name = "b",
  base = c("GGE", "EEE"),
  sugar = c("FFO", "OOO"),
  dictionary = my_dic
)
XNAStringSetObj <- XNAStringSet(objects = list(obj1, obj2, obj3))
}
\author{
Anna Gorska
}
