% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XNAString2Helm.R
\name{XNAStringToHelm}
\alias{XNAStringToHelm}
\title{XNAStringToHelmFun function takes XNAString object and translates
base, sugar and backbone to HELM notation}
\usage{
XNAStringToHelm(xnastring_obj, dictionary = xna_dictionary)
}
\arguments{
\item{xnastring_obj}{XNAString object}

\item{dictionary}{HELM-symbol dictionary}
}
\value{
string (HELM notation)
}
\description{
XNAStringToHelmFun function takes XNAString object and translates
base, sugar and backbone to HELM notation
}
\examples{
 obj <- XNAString(base = 'AAA',
                  sugar = 'DDD',
                  backbone = 'OO')
 XNAStringToHelm(obj)
}
